Separate component to render error. This will change in the future as design is updated
